+++ packages/scripts/src/bin.ts
@@ -1,7 +1,10 @@
 import chalk from 'chalk';
+import type { Configuration } from 'electron-builder';
+import { parse } from 'jsonc-parser';
 import assert from 'node:assert';
 import child from 'node:child_process';
-import fs from 'node:fs';
+import fs from 'node:fs/promises';
+import path from 'node:path';
 
 import { paths } from './utils/index.js';
 
@@ -27,15 +30,25 @@ export function appMain(): void {
 }
 
 export async function mkdirWorking(): Promise<void> {
-    return fs.promises
+    return fs
         .mkdir(paths.working)
         .then(() => console.info(`Create directory "${paths.working}"`))
         .catch(() => console.info(`Directory "${paths.working}" exists.`));
 }
 
-export const runBuild = () => child.spawn(paths.electron, [paths.repository], { cwd: paths.working, stdio: 'inherit' });
+export function runBuild() {
+    child.spawn(paths.electron, [paths.repository], { cwd: paths.working, stdio: 'inherit' });
+}
 
-export const runUnpacked = () => child.spawn(paths.unpacked, [], { cwd: paths.working, stdio: 'inherit' });
+export async function runUnpacked() {
+    const configRaw = await fs.readFile(path.join(paths.repository, 'electron-builder.json'), 'utf-8');
+    const { productName } = parse(configRaw) as Configuration;
+    if (productName) {
+        child.spawn(path.join(paths.unpackedWinDir, productName), [], { cwd: paths.working, stdio: 'inherit' });
+    } else {
+        console.error(chalk.red('Cannot find productName in electron-builder.json'));
+    }
+}
 
 export function unitTest(watch: boolean): void {
     const argv = process.argv.slice(2);